Tests/LibWeb/Text/expected/css/CSSImportRule-supportsText.txt

cssText: @import url("https://something.invalid/") supports(foo: bar);
supportsText: foo: bar
cssText: @import url("https://something.invalid/") supports((display: block) and (foo: bar));
supportsText: (display: block) and (foo: bar)
